automatically publish new package and create release on tag
