+++ scripts/travis.sh
@@ -27,11 +27,11 @@ elif [ "${TRAVIS_MODE}" = "funcTests" ]; then
 	if [ ${n} = ${maxRetries} ]; then
 		exit 1
 	fi
-elif [ "${TRAVIS_MODE}" = "releaseCanary" ]; then
+elif [ "${TRAVIS_MODE}" = "release" ] || [ "${TRAVIS_MODE}" = "releaseCanary" ]; then
   # update the version
   # make sure everything is fetched https://github.com/travis-ci/travis-ci/issues/3412
   git fetch --unshallow
-  node ./scripts/set-canary-version.js
+  node ./scripts/set-package-version.js
   if [[ $(node ./scripts/check-already-published.js) = "not published" ]]; then
     npm run lint
     npm run build
@@ -39,12 +39,17 @@ elif [ "${TRAVIS_MODE}" = "releaseCanary" ]; then
     # write the token to config
     # see https://docs.npmjs.com/private-modules/ci-server-config
     echo "//registry.npmjs.org/:_authToken=${NPM_TOKEN}" >> .npmrc
-    npm publish --tag canary
-    echo "Published canary."
-    curl https://purge.jsdelivr.net/npm/hls.js@canary
-    echo "Cleared jsdelivr cache."
+    if [ "${TRAVIS_MODE}" = "releaseCanary" ]; then
+      npm publish --tag canary
+      echo "Published canary."
+      curl https://purge.jsdelivr.net/npm/hls.js@canary
+      echo "Cleared jsdelivr cache."
+    elif [ "${TRAVIS_MODE}" = "release" ]; then
+      npm publish
+      echo "Published."
+    fi
   else
-    echo "Canary already published."
+    echo "Already published."
   fi
 else
 	echo "Unknown travis mode: ${TRAVIS_MODE}" 1>&2
